+++ src/components/Map.vue
@@ -1,12 +1,12 @@
 <template>
+  <!-- v-if="!$q.screen.xs" -->
   <NaverMap
-    v-if="!$q.screen.xs"
-    style="width: 75%; height: 100vh; float: right"
+    :class="isMobile ? 'isMobile' : 'isPc'"
     :mapOptions="MAP_OPTIONS"
     @onLoad="onLoadMap($event)"
   >
     <span v-if="markerData" v-for="marker in markerData" v-bind:key="marker._id">
-      <NaverMarker :latitude="marker.lat" :longitude="marker.lng" @click="onClickMarker(marker)"> </NaverMarker>
+      <NaverMarker :latitude="marker.lat" :longitude="marker.lng" @click="onClickMarker(marker, 'btn')"> </NaverMarker>
     </span>
 
     <NaverInfoWindow :marker="marker" :open="isOpen" @onLoad="onLoadInfoWindow($event)" id="info-window-all">
@@ -36,7 +36,7 @@
 </template>
 
 <script setup lang="ts">
-import {ref, computed, watch, Ref} from 'vue';
+import {ref, computed, watch, Ref, ComputedRef} from 'vue';
 import { useRouter } from 'vue-router';
 import { useQuasar } from 'quasar';
 
@@ -49,6 +49,8 @@ import {isEmpty} from 'lodash';
 const router = useRouter();
 const $q = useQuasar();
 const mapStore = useMapStore();
+
+const isMobile: ComputedRef<boolean> = computed(() => $q.screen.sm || $q.screen.xs);
 
 const markerData = computed(() => mapStore.markerData);
 
@@ -66,8 +68,6 @@ const isOpen = ref(true);
 const isEarly = ref(true);
 
 const selectedArchive: Ref<Archive | undefined> = ref();
-
-
 
 watch(markerData, () => {
   const firstMarker = markerData.value[0];
@@ -106,7 +106,10 @@ function onLoadMap(mapObject: unknown) {
   map.value = mapObject;
 }
 
-function onClickMarker(archive: Archive) {
+function onClickMarker(archive: Archive, type?: String) {
+  if(isMobile.value && type === 'btn') {
+    router.push(`/archive/${archive._id}`);
+  }
   // 카페 목록 상세 가져오기
   setSelectedArchive(archive);
 }
@@ -136,4 +139,16 @@ function onClickDetailBtn(id: string | undefined) {
 }
 </script>
 
-<style scoped></style>
+<style scoped>
+/* NaverMap */
+.isPc {
+  width: 75%;
+  height: 100%;
+  float: right
+}
+.isMobile {
+  width: 100%;
+  height: 100%;
+}
+
+</style>
