+++ src/layouts/MapLayout.vue
@@ -35,7 +35,9 @@
         <div id="troublemaker" class="bg-[#fff] h-screen mt-[-3.5rem] pt-14">
           <div class="h-screen">
             <!--overflow-y-hidden-->
-            <q-card class="relative h-screen py-4 my-card"> <slot /> </q-card>
+            <q-card :class="!isMobile ? 'relative h-screen py-4 my-card': ''">
+              <slot />
+            </q-card>
 
             <Map />
           </div>
